// src/models/investimentalRecord.ts
export class InvestimentalRecord {
    orderID: string;
    orderNumber: string;
    side: string;
    exchange: string;
    symbol: string;
    market: string;
    price: number;
    volume: number;
    disclosed: number;
    value: number;
    fee: number;
    term: string;
    validity: string;
    triggerType: string;
    triggerPrice: number;
    settlementTerm: string;
    settlementDate: string;
    settlementType: string;
    shortSell: string;
    accountId: string;
    accountName: string;
    lastTradeId: number;
    lastTradeTicket: number;
    status: string;
    initiatedBy: string;
    updatedBy: string;
    updateType: string;
    updateTime: string;
    requestId: string;
    requestType: string;
    requestStatus: string;
}